feat: 新增人员检测定时任务

+++ platform/src/main/java/com/seu/platform/dao/entity/PointStatistic.java
@@ -25,12 +25,12 @@ public class PointStatistic implements Serializable {
     /**
      * 超限点数
      */
-    private Integer thresholdExceeded;
+    private Double thresholdExceeded;
 
     /**
      * 未超限点数
      */
-    private Integer thresholdWithin;
+    private Double thresholdWithin;
 
     /**
      * 停机时间
@@ -45,12 +45,12 @@ public class PointStatistic implements Serializable {
     /**
      * 刷新正常
      */
-    private Integer normalRefresh;
+    private Double normalRefresh;
 
     /**
      * 刷新异常
      */
-    private Integer exceptionRefresh;
+    private Double exceptionRefresh;
 
     /**
      * 创建时间
